split packages: Move ArcSDE factories to org.geotools.arcsde.data
