+++ modules/plugin/arcsde/datastore/src/main/resources/META-INF/services/org.geotools.coverage.grid.io.GridFormatFactorySpi
@@ -1,1 +1,1 @@
-org.geotools.arcsde.ArcSDERasterFormatFactory
+org.geotools.arcsde.data.ArcSDERasterFormatFactory
