+++ modules/plugin/arcsde/datastore/src/main/resources/META-INF/services/org.geotools.data.DataStoreFactorySpi
@@ -1,2 +1,2 @@
-org.geotools.arcsde.ArcSDEDataStoreFactory
-org.geotools.arcsde.ArcSDEJNDIDataStoreFactory
+org.geotools.arcsde.data.ArcSDEDataStoreFactory
+org.geotools.arcsde.data.ArcSDEJNDIDataStoreFactory
