+++ modules/plugin/arcsde/datastore/src/test/java/org/geotools/arcsde/ArcSDEDataStoreFactoryTest.java
@@ -32,10 +32,8 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
-import org.geotools.arcsde.data.ArcSDEDataStore;
-import org.geotools.arcsde.data.ArcSDEDataStoreConfig;
-import org.geotools.arcsde.data.InProcessViewSupportTestData;
-import org.geotools.arcsde.data.TestData;
+
+import org.geotools.arcsde.data.*;
 import org.geotools.arcsde.session.ArcSDEConnectionConfig;
 import org.geotools.arcsde.session.ISession;
 import org.geotools.arcsde.session.UnavailableConnectionException;
@@ -144,7 +142,7 @@ public class ArcSDEDataStoreFactoryTest {
 
     /**
      * Test method for {@link
-     * org.geotools.arcsde.ArcSDEDataStoreFactory#createNewDataStore(java.util.Map)}.
+     * ArcSDEDataStoreFactory#createNewDataStore(java.util.Map)}.
      */
     @SuppressWarnings("unchecked")
     @Test
@@ -158,7 +156,7 @@ public class ArcSDEDataStoreFactoryTest {
     }
 
     /**
-     * Test method for {@link org.geotools.arcsde.ArcSDEDataStoreFactory#canProcess(java.util.Map)}.
+     * Test method for {@link ArcSDEDataStoreFactory#canProcess(java.util.Map)}.
      */
     @Test
     public void testCanProcess() {
@@ -169,7 +167,7 @@ public class ArcSDEDataStoreFactoryTest {
 
     /**
      * Test method for {@link
-     * org.geotools.arcsde.ArcSDEDataStoreFactory#createDataStore(java.util.Map)}.
+     * ArcSDEDataStoreFactory#createDataStore(java.util.Map)}.
      *
      * @throws IOException
      */
@@ -189,7 +187,7 @@ public class ArcSDEDataStoreFactoryTest {
 
     /**
      * Test method for {@link
-     * org.geotools.arcsde.ArcSDEDataStoreFactory#createDataStore(java.util.Map)}.
+     * ArcSDEDataStoreFactory#createDataStore(java.util.Map)}.
      *
      * @throws IOException
      * @throws SeException
