+++ modules/plugin/arcsde/datastore/src/test/java/org/geotools/arcsde/data/TestData.java
@@ -40,7 +40,7 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.TimeZone;
 import java.util.logging.Logger;
-import org.geotools.arcsde.ArcSDEDataStoreFactory;
+
 import org.geotools.arcsde.ArcSdeException;
 import org.geotools.arcsde.session.Command;
 import org.geotools.arcsde.session.Commands;
